# docs/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
  common.DataObject:
    properties:
      data:
        type: object
      id:
        type: string
      object_id:
        type: string
      token:
        type: string
    type: object
  common.GeoPoint:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
  dialog.Action:
    properties:
      action_id:
        type: string
      handable:
        type: boolean
      title:
        type: string
    type: object
  dialog.Error:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  dialog.Meta:
    properties:
      client_id:
        type: string
      locale:
        type: string
      location:
        $ref: '#/definitions/common.GeoPoint'
        type: object
      time_zone:
        type: string
      user_id:
        example: 009dew4enew1711123
        type: string
    type: object
  dialog.Request:
    properties:
      action_id:
        example: "yes"
        type: string
      client_entities:
        type: object
      meta:
        $ref: '#/definitions/dialog.Meta'
        type: object
      session:
        $ref: '#/definitions/dialog.Session'
        type: object
      user_response:
        example: Москва
        type: string
      version:
        type: string
    type: object
  dialog.Response:
    properties:
      actions:
        items:
          $ref: '#/definitions/dialog.Action'
        type: array
      error:
        $ref: '#/definitions/dialog.Error'
        type: object
      extras:
        type: object
      hint:
        type: string
      objects:
        items:
          $ref: '#/definitions/common.DataObject'
        type: array
      session:
        $ref: '#/definitions/dialog.Session'
        type: object
    type: object
  dialog.Session:
    properties:
      message_date_time:
        example: "2019-08-08T14:15:22+03:00"
        type: string
      message_id:
        type: integer
      session_id:
        example: a6e6f1c8-5f61-43c8-b9ee-f1dd176108cf
        type: string
    type: object
  places.Place:
    properties:
      cityName:
        type: string
      countryName:
        type: string
      description:
        type: string
      fullName:
        type: string
      id:
        type: integer
      lat:
        type: number
      lon:
        type: number
      name:
        type: string
      nearestBiggerPlaceId:
        type: integer
      placeclassId:
        type: integer
      platformName:
        type: string
      stateName:
        type: string
      stationName:
        type: string
      timeZone:
        type: string
      typePlace:
        type: integer
    type: object
  router.WebPlacesAutoSuggestionsRes:
    properties:
      data:
        properties:
          popular:
            items:
              $ref: '#/definitions/places.Place'
            type: array
          recent:
            items:
              $ref: '#/definitions/places.Place'
            type: array
        type: object
    type: object
  user.AuthorizeAndCompleteRegistrationUsecaseResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      has_email:
        type: boolean
      refresh_token:
        type: string
      request_id:
        type: string
      user_id:
        type: string
    type: object
host: '{{.Host}}'
info:
  contact: {}
  description: '{{.Description}}'
  license: {}
  title: maasapi
  version: '{{.Version}}'
paths:
  /api/auth/authorize:
    post:
      consumes:
      - application/json
      description: authorize
      operationId: authorize
      parameters:
      - description: Device
        in: body
        name: req
        schema:
          $ref: '#/definitions/AuthorizeReq'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.AuthorizeAndCompleteRegistrationUsecaseResponse'
            type: object
      summary: authAuthorize
      tags:
      - auth
  /api/auth/create:
    post:
      consumes:
      - application/json
      description: create temporary user
      operationId: create-temporary-user
      parameters:
      - description: Device
        in: body
        name: req
        schema:
          $ref: '#/definitions/CreateTemporaryUserReq'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CreateTemporaryUserRes'
            type: object
      summary: authCreateTemporaryUser
      tags:
      - auth
  /api/auth/sendSms:
    post:
      consumes:
      - application/json
      description: |-
        send sms code
        Demo user phone is +70001112233
      operationId: send-sms-code
      parameters:
      - description: Device
        in: body
        name: req
        schema:
          $ref: '#/definitions/SendSmsCodeReq'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SendSmsCodeRes'
            type: object
      summary: authSendSmsCode
      tags:
      - auth
  /api/dialogs/dialog:
    post:
      consumes:
      - application/json
      description: Dialog
      operationId: dialog
      parameters:
      - description: Dialog
        in: body
        name: req
        schema:
          $ref: '#/definitions/dialog.Request'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dialog.Response'
            type: object
      summary: Dialog
      tags:
      - dialog
  /api/places/autosuggestions:
    get:
      consumes:
      - application/json
      description: return recent and popular places
      operationId: web-places-auto-suggestions
      parameters:
      - description: user_id to get recent, if empty - recent is []
        in: header
        name: AS-CID
        type: string
      - description: direction
        enum:
        - from
        - to
        in: query
        name: direction
        required: true
        type: string
      - default: 5
        description: recent count in response
        in: query
        name: recentCount
        type: integer
      - default: 5
        description: popular count in response
        in: query
        name: popularCount
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.WebPlacesAutoSuggestionsRes'
            type: object
      summary: webPlacesAutoSuggestions
      tags:
      - web
swagger: "2.0"